GCP resources reconciled
